# user-management-api/public/swagger/swagger.json
{"swagger":"2.0","info":{"title":"User Management API","version":"1.0.0","description":"Manage Users"},"paths":{"/tokens":{"get":{"summary":"Get Token","description":"Get token for a given user credential","produces":["application/json"],"parameters":[{"name":"username","in":"query","description":"username","required":true,"type":"string"},{"name":"userPassword","in":"query","description":"password","required":true,"type":"string"}],"tags":["Tokens"],"responses":{"200":{"description":"Successful operation","schema":{"$ref":"#/definitions/TokenResponse"}}}},"post":{"summary":"Get token","description":"Get token for a given user credential","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"userCredential","in":"body","description":"user credential","required":true,"schema":{"$ref":"#/definitions/TokenRequest"}}],"tags":["Tokens"],"responses":{"200":{"description":"Successful operation","schema":{"$ref":"#/definitions/TokenResponse"}},"400":{"description":"user Validation Error.","schema":{"$ref":"#/definitions/Response"}}}}},"/users/{username}":{"get":{"summary":"Get User by username","description":"get User","produces":["application/json"],"parameters":[{"name":"username","in":"path","description":"username","required":true,"type":"string"}],"tags":["Users"],"responses":{"200":{"description":"Successful operation","schema":{"$ref":"#/definitions/UserResponse"}},"401":{"description":"Unauthorized request","schema":{"$ref":"#/definitions/Response"}}}},"delete":{"summary":"Delete user","description":"delete User","produces":["application/json"],"parameters":[{"name":"username","in":"path","description":"username","required":true,"type":"string"}],"tags":["Users"],"responses":{"200":{"description":"Successful operation","schema":{"$ref":"#/definitions/Response"}},"401":{"description":"Unauthorized request","schema":{"$ref":"#/definitions/Response"}}}},"put":{"summary":"Update User","description":"update User","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"username","in":"path","description":"username","required":true,"type":"string"},{"name":"user","in":"body","description":"user entity","required":true,"schema":{"$ref":"#/definitions/User"}}],"tags":["Users"],"responses":{"200":{"description":"Successful operation","schema":{"$ref":"#/definitions/UserResponse"}},"400":{"description":"user Validation Error.","schema":{"$ref":"#/definitions/Response"}},"401":{"description":"Unauthorized request","schema":{"$ref":"#/definitions/Response"}}}}},"/users":{"get":{"summary":"Get current User","description":"Get current User","produces":["application/json"],"parameters":[{"name":"email","in":"query","description":"email","required":false,"type":"string"},{"name":"username","in":"query","description":"username","required":false,"type":"string"},{"name":"lastName","in":"query","description":"last name","required":false,"type":"string"},{"name":"firstName","in":"query","description":"first name","required":false,"type":"string"},{"name":"city","in":"query","description":"city","required":false,"type":"string"},{"name":"state","in":"query","description":"state","required":false,"type":"string"},{"name":"postalCode","in":"query","description":"Postal Code","required":false,"type":"string"}],"tags":["Users"],"responses":{"200":{"description":"Successful operation","schema":{"$ref":"#/definitions/UserResponse"}},"401":{"description":"Unauthorized request","schema":{"$ref":"#/definitions/Response"}}}},"post":{"summary":"Create User","description":"Create User","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"user","in":"body","description":"user entity","required":true,"schema":{"$ref":"#/definitions/User"}}],"tags":["Users"],"responses":{"200":{"description":"Successful operation","schema":{"$ref":"#/definitions/UserResponse"}},"400":{"description":"user Validation Error.","schema":{"$ref":"#/definitions/Response"}},"401":{"description":"Unauthorized request","schema":{"$ref":"#/definitions/Response"}}}}}},"basePath":"/user-management/api","definitions":{"Outcome":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"errors":{"type":"array","items":{"$ref":"#/definitions/ReturnStatus"}}}},"Response":{"type":"object","properties":{"outcome":{"type":"object","$ref":"#/definitions/Outcome"},"data":{"type":"object"}}},"ReturnStatus":{"type":"object","properties":{"message":{"type":"string"},"name":{"type":"string"}}},"User":{"type":"object","properties":{"username":{"type":"string"},"userPassword":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"},"email":{"type":"string"},"addressLine1":{"type":"string"},"addressLine2":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"postalCode":{"type":"string"},"homePhone":{"type":"string"},"workPhone":{"type":"string"},"mobilePhone":{"type":"string"}}},"Token":{"type":"object","properties":{"token":{"type":"string"},"username":{"type":"string"}}},"TokenResponse":{"type":"object","properties":{"outcome":{"type":"object","$ref":"#/definitions/Outcome"},"data":{"type":"array","items":{"$ref":"#/definitions/Token"}}}},"TokenRequest":{"type":"object","properties":{"username":{"type":"string"},"userPassword":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"outcome":{"type":"object","$ref":"#/definitions/Outcome"},"data":{"type":"array","items":{"$ref":"#/definitions/User"}}}}}}